fix: render page titles from frontmatter in browser tab
